backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
billiard==4.2.4
blinker==1.9.0
boltons==21.0.0
bracex==2.6
CacheControl==0.14.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
click-didyoumean==0.3.1
click-option-group==0.5.9
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.3.1
distro==1.9.0
exceptiongroup==1.2.2
face==26.0.0
fastapi==0.135.1
firebase_admin==7.2.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==4.1.1
glom==25.12.0
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-core==2.5.0
google-cloud-firestore==2.24.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.23.3
mdurl==0.1.2
msgpack==1.1.2
numpy==2.4.2
openai==2.26.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
ordered-set==4.1.0
packaging==26.0
peewee==3.19.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.14
semantic-version==2.10.0
semgrep==1.154.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sse-starlette==3.3.2
starlette==0.52.1
tomli==2.0.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcmatch==8.5.2
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==1.17.3
zipp==3.23.0
